Add files via upload
Game feature: 
The sun has three life. The sun should have one  small red sphere gun at its outer edge.  The gun should shoot one red colored bullet from the red sphere on the outer edge when we right click on the sun.  If the bullet hits any of the planets the sun will have an increased score . This score will be displayed on the screen . If the 3 consecutive bullets miss all of the planets one life will be reduced. The life count, bullets missed will also be shown on the screen . If the life becomes 0 screen should show game over and score on the screen. Then if we press r it should restart the game.